refactor: 경로 정리 및 통합
