+++ worker/src/application/ports/receiver.ts
@@ -1,4 +1,4 @@
-import { EventDetail } from "application/dto";
+import { EventDetail } from "../dto";
 
 export interface IEventReceiver {
     receive(params: any): Promise<EventDetail>;
