+++ worker/src/application/ports/sender.ts
@@ -1,4 +1,4 @@
-import { EventDetail } from "application/dto";
+import { EventDetail } from "../dto";
 
 export interface INotificationSender {
     dispatch(notification: EventDetail): Promise<void>;
