+++ worker/src/infrastructure/api/index.ts
@@ -1,0 +1,2 @@
+export { CalendarClient } from "./calendarClient";
+export { OnesignalClient } from "./onesignalClient";
