+++ worker/src/module.ts
@@ -3,13 +3,12 @@ import { ConfigModule } from "@nestjs/config";
 import { ScheduleModule } from "@nestjs/schedule";
 
 import { WorkerService } from "application/usecases/service";
+import { WorkerCronService } from "application/usecases/cron";
 
-import { WorkerCronService } from "infrastructure/cron";
 import { WorkerClientImpl } from "infrastructure/clientImpl";
-import { CalendarClient } from "infrastructure/api/calendarClient";
-import { OnesignalClient } from "infrastructure/api/onesignalClient";
 import { WebNotificationSender } from "infrastructure/senders/webSender";
 import { CalendarEventReceiver } from "infrastructure/receivers/calendarReceiver";
+import { OnesignalClient, CalendarClient } from "infrastructure/api";
 
 @Module({
     imports: [ScheduleModule.forRoot(), ConfigModule.forRoot()],
